Use pragma once
The "once" pragma should be supported everywhere that the ## pre-processor statement is (see https://en.wikipedia.org/wiki/Pragma_once for the list of supported compilers). 

Apart from less code in the header file, it also solves naming collisions. LIST_H may be a bit generic and inviting a collision from elsewhere in a large project.